+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Improve commit behavior
 - Move gg commit conventions from gg_git to gg_one_core
+- Record the doCommit state in system commits again
 
 ## 2.1.0 - 2026-08-09